+++ googleapis_auth/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added parsing for project_id and universe_domain properties for ServiceAccountCredentials
 - Add `sign()` method to `ServiceAccountCredentials` for RSA-SHA256 signing
 - Add `IAMSigner` class for signing via IAM Credentials API
+- Add `clientViaServiceAccountImpersonation()` function and `ImpersonatedAuthClient` class for service account impersonation via IAM Credentials API
 - Require `meta: ^1.0.2`
 - Require `sdk: ^3.9.0`
 - Drop unneeded `args` dependency.